Fix lint and mypy issues in resource + urban-composite work
Accumulated across the 8 unpushed commits since 2e66aad — last
green CI run. None of these changed CI status because nothing has
been pushed since, but the next push would have failed all three
checks (ruff check, ruff format, mypy).

- asset_index.py: inline negated EVENTPACK_ gate (SIM103)
- clutter_culling.py: drop unused PIL.Image import; cast
  np.linalg.inv result to float64 so it matches the declared
  list[tuple[NDArray[uint8], NDArray[float64]]] (mypy)
- tests: rename unused lambda args to _-prefixed (ARG005), add
  strict=True to a length-asserted zip (B905), promote the
  ClutterMaskPart import to module level so the return annotation
  resolves without a type: ignore (F821 / UP037)
- extractor.py + three test files: ruff format whitespace

No behavior change; tests still 293/293 pass.

+++ src/pinacotheca/clutter_culling.py
@@ -31,7 +31,6 @@ from typing import Any
 
 import numpy as np
 from numpy.typing import NDArray
-from PIL import Image
 
 from pinacotheca.clutter_transforms import TERRAIN_CLUTTER_TYPE_NONE
 from pinacotheca.prefab import PrefabPart
@@ -179,7 +178,7 @@ def cull_clutter_against_masks(
         if img is None:
             continue
         try:
-            inv = np.linalg.inv(plane.world_matrix)
+            inv = np.linalg.inv(plane.world_matrix).astype(np.float64)
         except np.linalg.LinAlgError:
             logger.warning(
                 "TerrainClutterSplat %r world matrix is singular; skipping",
